refactor: Added functions for calculating the canvas width and height.
